Moved ServerInactivity & ServerWarning to Models.Server
