+++ TreeDiagram/Models/Server/ServerWarning.cs
@@ -2,7 +2,7 @@ using System.Collections.Generic;
 using System.Linq;
 using TreeDiagram.Models.SubModels;
 
-namespace TreeDiagram.Models.Server.Warning
+namespace TreeDiagram.Models.Server
 {
     public class ServerWarning : ConfigBase
     {
